+++ Pasteboard/Utilities/VisualConfiguration.swift
@@ -5,6 +5,7 @@ enum VisualConfiguration {
     static let panelSize = AppConfiguration.panelSize
     static let headerIconSize: CGFloat = 36
     static let headerSpacing: CGFloat = 10
+    static let panelTopPadding: CGFloat = 6
     static let rowSpacing: CGFloat = 8
     static let thumbnailSize = CGSize(width: 56, height: 56)
     static let searchFieldSpacing: CGFloat = 12